PM-38588 - Move CurrentContextBackfillService out of RequestValidators
It's a side-effect helper consumed by the grant validators, not a request
validator itself — same shape as UserDecryptionOptionsBuilder, which already
lives at the IdentityServer root next to CustomValidatorRequestContext.
